// src/utils/entities.type.ts
export enum eUserStatus {
  ACTIVE = 'active',
  DE_ACTIVE = 'de_active'
}

export enum eModule {
  ROLE_PERMISSION = 'role_permission',
  USER_MANAGEMENT = 'user_management',

}

export enum eModulePermission {
  VIEW = 'view',
  MANAGE = 'manage'
}


export enum eCoreTeamStatus {
  ACTIVE = 'active',
  OFF_BOARD = 'off_board'
}